+++ src/App.tsx
@@ -22,6 +22,7 @@ import Personnages from './screens/Personnages';
 import FichePersonnage from './screens/FichePersonnage';
 import Quiz from './screens/Quiz';
 import ModeIA from './screens/ModeIA';
+import AideIA from './screens/ia/AideIA';
 import Frise from './screens/Frise';
 import Carte from './screens/Carte';
 import Collections from './screens/Collections';
@@ -72,6 +73,7 @@ export default function App() {
           <Route path="/explorer/personnage/:id" element={<FichePersonnage />} />
           <Route path="/quiz" element={<Quiz />} />
           <Route path="/ia" element={<ModeIA />} />
+          <Route path="/ia/aide" element={<AideIA />} />
           <Route path="/explorer/frise" element={<Frise />} />
           <Route path="/explorer/carte" element={<Carte />} />
           <Route path="/explorer/themes" element={<Themes />} />